refactor(scope): rename ScopeBreadcrumb→ScopeHeader, inline ScopedTitle
ScopeBreadcrumb was a <header> wrapper, not the breadcrumb itself —
the name caused a confusing breadcrumb→ScopedTitle→Breadcrumb chain.
ScopedTitle had no other callers so it's inlined directly into ScopeHeader.
